lot of changes

+++ api/controllers/assignments.py
@@ -1,10 +1,10 @@
 from marshmallow import ValidationError
 from bson.objectid import ObjectId
 from flask import request, Blueprint, abort
-from api.auth import get_authenticated_user, auth_required, roles_required
+from api.dataaccess.auth import get_authenticated_user, auth_required, roles_required
 from api.extensions import mongo
 from api.responses import ok
-from api.models.assignments import create_assignment_schema, update_assignment_schema
+from api.models.assignment import create_assignment_schema, update_assignment_schema
 
 
 assignments_blueprint = Blueprint('/api/assignments', __name__)
@@ -29,8 +29,8 @@ def get_assignment(id):
 @auth_required
 def get_assignments():
     user = get_authenticated_user()
-    assignment = mongo.db.assignments.find({'user_id': user['_id']})
-    return ok(list(assignment))
+    assignments = mongo.db.assignments.find({'user_id': user['_id']})
+    return ok(list(assignments))
 
 
 @assignments_blueprint.route('/api/assignments', methods=['POST'])
@@ -39,7 +39,16 @@ def post_assignment():
     try:
         data = create_assignment_schema.load(request.get_json())
         user = get_authenticated_user()
+        
         data['user_id'] = user['_id']
+        data['course_id'] = ObjectId(data['course_id'])
+        course = mongo.db.courses.find_one({'_id': ObjectId(data['course_id'])})
+
+        if not course:
+            abort(400)
+
+        if 'admin' not in user['roles'] and course['user_id'] != user['_id']:
+            abort(403)
 
         result = mongo.db.assignments.insert_one(data)
         print(result)
@@ -72,10 +81,17 @@ def put_assignment():
 
 
 @assignments_blueprint.route('/api/assignments/<id>', methods=['DELETE'])
-@roles_required(['admin'])
+@auth_required
 def delete_user(id):
-    resp = mongo.db.assignments.delete_one({'_id': ObjectId(id)})
-    if resp.deleted_count:
-        return ok()
-    else:
+    assignment = mongo.db.assignments.find_one({'_id': ObjectId(id)})
+
+    if not assignment:
         abort(404)
+
+    user = get_authenticated_user()
+    if 'admin' not in user['roles'] and assignment['user_id'] != user['_id']:
+        abort(403)
+
+    resp = mongo.db.assignments.delete_one({'_id': ObjectId(id)})
+
+    return ok()
